+++ tests/linalg_test.py
@@ -1683,10 +1683,13 @@ class ScipyLinalgTest(jtu.JaxTestCase):
     mode=["full", "r", "economic"],
     pivoting=[False, True]
   )
+  @jax.default_matmul_precision("float32")
   def testScipyQrModes(self, shape, dtype, mode, pivoting):
-    is_not_cpu_test_device = not jtu.test_device_matches(["cpu"])
-    if pivoting and is_not_cpu_test_device:
-      self.skipTest("Pivoting is only supported on CPU with jaxlib > 0.4.38")
+    if pivoting:
+      if not jtu.test_device_matches(["cpu", "gpu"]):
+        self.skipTest("Pivoting is only supported on CPU and GPU.")
+      if jtu.test_device_matches(["gpu"]) and jtu.jaxlib_version() <= (0, 5, 0):
+        self.skipTest("Pivoting is only supported on GPU for jaxlib > 0.5.0")
     rng = jtu.rand_default(self.rng())
     jsp_func = partial(jax.scipy.linalg.qr, mode=mode, pivoting=pivoting)
     sp_func = partial(scipy.linalg.qr, mode=mode, pivoting=pivoting)
@@ -1700,7 +1703,7 @@ class ScipyLinalgTest(jtu.JaxTestCase):
     def qr_and_mul(a):
       q, r, *p = jsp_func(a)
       # To express the identity function we must "undo" the pivoting of `q @ r`.
-      inverted_pivots = p[0][p[0]]
+      inverted_pivots = jnp.argsort(p[0])
       return (q @ r)[:, inverted_pivots]
 
     m, n = shape